requirements.txt
# Trading Bot Dependencies
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.1.0
python-dateutil==2.8.2
werkzeug==3.0.1

# Logging and Monitoring
colorlog==6.8.0

# Data validation
jsonschema==4.20.0

